+++ lib/src/socket.dart
@@ -1,0 +1,8 @@
+import 'package:meta/meta.dart';
+
+/// An interface to a connection to a remote party.
+@sealed
+abstract class Socket {
+  /// Creates an instance of `Socket`.
+  const Socket();
+}